Reserve Client enum values for testing and remove an unused client mapping.
Internal testing requires the addition of `Client` enum values that will not be
included or used in the public APK. `SUSPICIOUS_MESSAGE_ALERTS` is no longer
used and can be removed from the mapping.

This change also updates the network logger pattern for
`PLAY_PROTECT_SERVICE_CORE_DEFAULT` client ids.

PiperOrigin-RevId: 561379562